Revamping Logic
Lots more info in the comments of the code.

Documentation still needs to be made, also more testing lol

This is the new format for Logic.json files:

```
{
    "Rule-1":{
        "IF": [
            "Variant collection name"
        ],
        "THEN":[
            "Attribute collection name", "Variant collection name"
        ]
    }
}
```

Some basic guidelines, needs to be more thorough and cleaned up for documentation:

Attributes and Variants can be used in the “THEN” list, Variants can only be used in the “IF” list.

If a full attribute is detected in the “THEN” list (either by the attribute collection name or all variant collection names in that attribute) and a variant is not detected in the “IF” list, the attribute detected in the “THEN” list will be set to “EMPTY”.

Variants in the “IF” list may be from the same or separate attributes, however they must never be from the same attribute as variants in the “THEN” list.

I need sleep.

Note to self: The Logic UI List needs to be updated to reflect these changes to Logics core functionality, the current UI List is broken on this branch and wont work.

Also added save_result to Constants.py for basic logging functionality.